[v13] Clean up Drone slack notifcations (#25216)
* Pin the plugins/slack image for dronegen pipelines

The latest version of this image has a regression where it reports
builds as successful even when they've failed.

* Rework slack notifications in dronegen

Remove the call to action language, and leave only the relevant links.

Flatten several short lines into one longer line.  This allows the
entire message to render without needing to expand it.

* Pin the plugins/slack image for unmanaged pipelines

The latest version of this image has a regression where it reports
builds as successful even when they've failed.

* Rework slack notifications in unmanaged pipelines

Remove the call to action language, and leave only the relevant links.

Flatten several short lines into one longer line.  This allows the
entire message to render without needing to expand it.