Create OktaAssignments for users on login. (#1198)
When a user logs in, OktaAssignments will be created for that user based on
their permissions and what Okta objects they have access to. This will then
be processed by the assignment reconciliator as normal.

When a new Okta assignment is created for a user, the old one will be marked
as CLEANED_UP. The reconciliator should then be able to make sense of both
to ensure that access is not lost to the user.

Source-commit: gravitational/teleport.e@0ca3b1c8f4713dc3b4626571b8f84c170132c228
Source-pull-requests: gravitational/teleport.e#1198