[v10] Clean up Drone slack notifcations (#25220)
* dronegen: Fix nesting of slack template parameter

When generating a GitHub Actions pipeline, put the slack message
template under `step.settings` instead of `step`, as the template is a
parameter to the plugin, not a configuration item of the step.

The `Template` field (now unused) has not been removed from the `step`
struct as it needs to be there to unmarshal the existing `.drone.yml`
file. Once that has been regenerated, we can remove the field.

* drone: Regenerate .drone.yml with correct templates

Regenerate `.drone.yml` with the correct template setting for the slack
plugin.

* dronegen: Remove unused Template field from step

Remove the unused `step.Template` field as it is no longer used. It
could not be removed earlier as it was needed to be able to unmarshal
the existing `.drone.yml` file. Now that has been done in a previous
commit, the field can be removed.

* Pin the plugins/slack image for dronegen pipelines

The latest version of this image has a regression where it reports
builds as successful even when they've failed.

* Rework slack notifications in dronegen

Remove the call to action language, and leave only the relevant links.

Flatten several short lines into one longer line.  This allows the
entire message to render without needing to expand it.

* Pin the plugins/slack image for unmanaged pipelines

The latest version of this image has a regression where it reports
builds as successful even when they've failed.

* Rework slack notifications in unmanaged pipelines

Remove the call to action language, and leave only the relevant links.

Flatten several short lines into one longer line.  This allows the
entire message to render without needing to expand it.

---------

Co-authored-by: Cam Hutchison <camh@goteleport.com>