[v12] build: Copy release artifacts to release directory
Copy the tarball and mac packages/dmgs release artifacts to a release
artifact directory so that the CI scripts do not have to. This makes it
clearer what is and is not a release artifact and puts the logic in the
Makefile instead of the CI yaml, so it can more easily be tested locally
and to make it easier to migrate to the next CI system.