Fix `TestKubernetesLicenseEnforcement` flakiness
When `TestKubernetesLicenseEnforcement` parallel tests run, they inherit
the  `err` var from the main function and tried to write the result into
it without any protection. This resulted in a data race detected.

The PR allocates the `err` var locally to the closure in order to
prevent the data race.

Fixes #25357