Allow `RoleRemoteProxy` to list 'kube_servers` when no Kube license (#25492)
When a remote proxy retrieves the list of kube clusters available in the
remote cluster and the remote cluster isn't licensed, the request fails because the remote proxy role was not
whitelisted. This PR does not return the license error if the requester is a local or remote server.

This PR also disables the `kubernetes_service` on startup if the cluster is
not licensed for Kubernetes access.

This prevents the process from being in the degraded state when trying to
upsert the `kube_server` via heartbeat.