build: Fix ELECTRON_BUILDER_ARCH to not throw error (#25676)
Do not cause a `make` `$(error ...)` if the arch is not a known electron
builder architecture. A previous commit should have added

    unexport ELECTRON_BUILDER_ARCH

but it was forgotten. Instead of adding that, just add mappings where
needed and let `$(ARCH)` map directly to an electron builder arch flag
when there is no mapping. If this is invalid for an architecture we
want to build for, let EB complain then.