add redirect to windows user creation instructions to host user creation doc
I feel it's natural for users to end up on host user creation when searching for Windows user creation docs.  Adding a redirect at the beginning for those searching for Windows instructions, especially since those instructions are not a standalone page like SSH but a subsection of the Windows RBAC guide.