WIP: clean up moderated sessions RBAC
Most of the changes here are an attempt to massage our existing
resource RBAC checker into handling session joining as a special
case.

After a bit of debugging, I'm thinking the better approach will be
to have a separate RBAC entry point for session joining. This will
allow us to avoid having a bunch of special cases littered througout
the code.