apiserver.go: Remove Auth Failure Logging
With the broad and increased rate limits we are at times "optimistically" checking the auth.
That has introduced noise in the logs in cases where optimism fails us.

It does not sound like this logging is adding much debug value, and we wanted to avoid adding complexity through conditional logging.
For that reason this commit removes the logging for all auth failure cases.