ci: Always save arch-specific Mac build artifacts (#1329)
Make sure we always save the artifacts from the architecture-specific
builds as they are needed as inputs to the universal build workflow.
These will still not be uploaded if `release-artifacts` is false.

This broke the push build which runs with `release-artifacts = false`
since we are not doing a release. This was not noticed on master as we
do not do push builds at the moment.

Source-commit: gravitational/teleport.e@845d4ffd05b643fc31ee872e0bbd2fa0879966c5
Source-pull-requests: gravitational/teleport.e#1329