build: Scope RUST_VERSION var to single target (#25960)
Scope the `RUST_VERSION` variable to the single
`rustup-install-target-toolchain` target that needs it. Setting it at
the top-level causes a strange interaction between the top-level
Makefile and the one in `build.assets/` that causes `make fix-imports`
(and presumably any other target that chains to `build.assets/Makefile`
and uses `RUST_VERSION`) to get an empty `RUST_VERSION`.

We use `:=` when setting `RUST_VERSION` otherwise the same problem is
present but just for the `rustup-install-target-toolchain` target.