[v12] add redirect to windows user creation instructions to host user creation doc (#25966)
* add redirect to windows user creation instructions to host user creation doc

I feel it's natural for users to end up on host user creation when searching for Windows user creation docs.  Adding a redirect at the beginning for those searching for Windows instructions, especially since those instructions are not a standalone page like SSH but a subsection of the Windows RBAC guide.

* Update docs/pages/server-access/guides/host-user-creation.mdx

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

* Update host-user-creation.mdx

* Update host-user-creation.mdx

* Update host-user-creation.mdx

* correct link

---------

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>
Co-authored-by: Steven Martin <steven@goteleport.com>