tctl: improve alert ack flows (#26029)
* Make --reason required for `tctl alert ack`

The backend requires a reason, so by having it optional in the CLI
it just means we have to make a network request in order to get
an error.

Also update the docs to indicate that this is not optional.

* tctl: show acknowledged alerts when the -v flag is provided

* Add a --format flag for tctl alerts ls

This makes it easier to programatically operate on the alerts.