ci: Add a checkout-teleport action (#1341)
Add an action - `checkout-teleport` - to checkout a teleport OSS repo
and optionally a teleport enterprise repository.

This factors out the five steps that the build-linux and build-mac
workflows have in common. There are some other workflows that also
checkout one or both repos that could also call this, but there are a
few little things to work out first. Baby steps.

The action currently lives in this teleport.e repo rather than as a
separate repository as is often done with actions. There may be some
security implications we want to assess before splitting this out if we
want to. For now, leaving it as a local action makes this a simple
refactor.

Source-commit: gravitational/teleport.e@fe2532f7d55a1bebbabfcfab05020dd868f138e3
Source-pull-requests: gravitational/teleport.e#1341