[v12] Use os.UserHomeDir where possible (#25998)
* User os.UserHomeDir where possible

In #24156 we removed a lot of unnecessary calls to user.Current,
which can take a long time in Active Directory environments.

This commit removes a few other uses of user.Current where we were
doing a full user lookup only to find the user's home directory,
which is usually more readily available via the environment.

* Fix tsh profile dir

This fixes a bug introduced in #25950