ci: Split MacOS architecture-specific and universal builds (#1381)
Split the `build-mac.yaml` workflow so that it now builds just the
architecture-specific tarballs and the new `build-mac-universal.yaml`
builds the universal release artifacts from that. Originally it was
planned to maintain the ability to release architecture-specific
`.pkg` and `.dmg` artifacts, but that is not required. Without that, the
workflows are simpler by being separate.

Source-commit: gravitational/teleport.e@a86b5d12b6c6d62bc5efa1ed9ff766760cc9f4d2
Source-pull-requests: gravitational/teleport.e#1381