Allow non moderated sessions when no-auth connection exists (#26074)
* Allow non moderated sessions when no-auth connection exists

This PR builds the last bit required to allow users to create remote
kubernetes sessions when Auth connection is not available or is
intermitent.

This PR allows a user to continue with the interactive session when the
session does not require any moderation.

PR #25202 deferred the proxy cert creation for cases where the it's
mandatory (kube agent or proxy running version <=12.x.x) which allowed
any user to perform simple requests against a Kubernetes cluster even
when the cluster Auth server is not operational. On top of that, this PR
allows a user to request an interactive session against a pod when auth
conectivity is not required as long as the session isn't moderated.

Part of #25541
Related to #25202
Implements #26014

* fix formatting