Make PAM user creation script copy/pasteable (#26277)
Closes #18343

The user creation script in our SSH PAM guide is inside a heredoc within
a `code` snippet. The shebang within the script does not render within
the `code` snippet, since the snippet component treats the hash as a
comment.

This change turns the script into a code snippet with a `text` label to
prevent rendering issues and separates commands that are intended to be
run outside the script into their own code snippets.