Add No-Auth Kubernetes integration test
This PR adds an integration test that validates if a user with a valid
certificate is able to request pods from a Teleport cluster when Auth
connection is unavailable.

It also validates that a user can create a interactive Kubernetes
sessions if he does not require any moderation. If moderation is
required, the session is rejected and an error is returned back to
the client. Both scenarios are covered by integration tests.

Part of #25541