Turn the `tsh` step into a Details box
Step 4 of the Linux Server guide, which instructs the user to log in to
a cluster via `tsh`, breaks the user's focus from the Web UI
unnecessarily. `tsh` isn't complex enough to require its own step in
this guide.

For users who would like CLI access to Teleport, this change includes
the information as a Details box.