Amazon RDS converter: extract Subnets (#26621)
When parsing Amazon RDS databases extract the Subnets.

Why do we need them?
We are going to allow for automatic deployment of Database Services
using Amazon ECS w/ Fargate.
To do so, we must provide the subnets which our container can connect
to.

So, we need to get the Subnets associated with the Amazon RDS database
in order to pass them when creating the Amazon ECS Service.

Related issue:
https://github.com/gravitational/teleport/issues/25521