Fix moderated sessions joining when Kube cluster has multi-agents
This PR fixes a bug introduced in the original Moderated
sessions implementation (#8563). When the same Kubernetes cluster is
served by multiple kubernetes agents, the joining function didn't
consider the `host_id` that owns the moderated session. This results in
users receiving `session not found` errors when the reversetunnel
dialer picks a random agent that does not own the session.

Given that we encourage multiple agents running in the same kube
cluster, a user has $1/n, where n=number of agent replicas$ probability
of being able to join the session.

This PR fixes this bug by forcing the dialer to filter out kube agents
that do not own the session. A new field, `HostID` was introduced into
the session tracker to store this information and only Kubernetes
proxies will request the session tracker from the Auth server to
retrieve that value. With `host_id`, the kube proxy is able to filter
the host that owns the session and will connect to it. If the agent is
no longer available, an error is returned to the user.

Fixes #26684