Add the Jamf service skeleton (#1438)
Add an initial iteration of the Jamf service. The service can be registered
using an "mdm" token and started. It'll, for now, list devices so we can be sure
the setup is correct.

The skeleton is a slight variation of the already existing Okta service.

Further iterations will add the sync logic to it.

#826

Source-commit: gravitational/teleport.e@9df1320124dedd071cd6c63faf86b71dce716118
Source-pull-requests: gravitational/teleport.e#1438