[v13] Make the Linux Server guide less SSH-centric (#26631)
* Make the Linux Server guide less SSH-centric

The current Linux Server guide includes instructions for accessing the
server and viewing a session recording. For users who are not interested
in server access, this can be irrelevant or even a turnoff. Instead of
showing SSH-specific features, make the final step about registering
resources using the Web UI.

* Turn the `tsh` step into a Details box

Step 4 of the Linux Server guide, which instructs the user to log in to
a cluster via `tsh`, breaks the user's focus from the Web UI
unnecessarily. `tsh` isn't complex enough to require its own step in
this guide.

For users who would like CLI access to Teleport, this change includes
the information as a Details box.

* Tweak the language of the "Add Resources" step

Use the term "enroll" instead.

* Partially respond to alexfornuto feedback

- Remove image
- Change language

* Use light mode for images