Fix port forwarding when using a label based target (#26394) (#26704)
Label based connections that do not provide a command and open a
shell for the first match were only performing a portion of the logic
that non label based interactive sessions were. Instead of connecting
directly to a host and opening a session, `tc.runShellOrCommandOnSingleNode`
is called to achieve the same behavior as non label based connections.

In addition to fixing #23261 this should also add support for `-N` in
the same scenario.