Fix link to CA Pinning information (#26690)
Closes #26666

The "CA Pinning" link in the CA rotation guide links to an H3 heading
within a closed `Details` box, making the link appear broken.

While it is possible to link to a `Details` box using a URL fragment,
the docs linter currently flags these links as pointing to a nonexistent
heading, since it is not yet aware of the `Details` box navigation
logic.

As a result, this change removes this link, describes CA pinning in
prose, and links to the general join token guide.