tooling: Add flakiness mode to render-tests
Add a `-report-by flakiness` mode to `render-tests` that can accumulate
multiple test runs and report the top N flaky tests. This mode is
intended to be used in a daily run of the tests for a couple of hours,
or perhaps 200 times, and to report on the tests that have failed the
most.

The output of multiple test runs should be fed into the single run of
`render-tests`. A `rerun` utility is forthcoming with which you could
do:

    rerun -n 200 -t 2h go test -shuffle on -cover -json . | \
        render-tests -report-by flakiness -top 10