Clean up LDAP error handling (#26984)
The following error often confuses users:

    *ldap.Error LDAP Result Code 1 &#34;Operations Error&#34;: 000004DC: LdapErr: DSID-0C090ACD,
    comment: In order to perform this operation a successful bind must be completed on the connection.

Since Teleport always uses x509 certs to bind, this error indicates
that the Teleport-issued cert is not trusted (which is likely due to
Teleport's CA not being imported as a trusted root).

While fixing this, unify the LDAP error handling with a common
utility for converting between LDAP error codes and trace errors.