Fetch ClusterAlerts a single time during login (#26903)
`onLogin` was inadvertently retrieving ClusterAlerts twice by
explicitly retrieving them and by calling `onStatus`, which also gets
the alerts so they are shown when `tsh status` is invoked. The portion
of `onStatus` which prints the profile has been separated into
`outputProfiles` so that `onLogin` may call that directly instead
of `onStatus`.