Add Okta import rule regex matching to synchronization. (#1446)
During synchronization, the Okta service will now take advantage of the new
regex matching features added to the Okta import rules. This will allow for
dynamic labeling of groups and applications rather than needing to list out
groups and applications one by one.

Prioritized labels are now calculated dynamically as a result, so references
to priorityAndLabels have been converted to pointers to try to reduce the
amount of copies of priorityAndLabel objects in memory.

The mutexes were removed for the existing import rule mappings, as they are
used in the synchronizer which is not run concurrently. This was adding
unnecessary complexity to the code.

Source-commit: gravitational/teleport.e@731db85f522f3a1b559340cd6a2ca18eefc04893
Source-pull-requests: gravitational/teleport.e#1446