feat: extend label expression syntax
This commit adds the new function `labels_matching` to the label
expression syntax. It accepts a Teleport-style glob or regular
expression, and returns the aggregate of all values for keys matching
that expression.

Since it is now more likely for users to want to compare two slices,
I've also added the `contains_any` and `contains_all` functions to
complement the existing `contains` function.

This commit also adds the `email.local` function that is already
available in role templates and should be useful.

Some examples are included in the new test cases.