Implement the MDM sync scheduler (#1554)
The MDM sync scheduler calculates the sync schedule, taking a list of inventory
entries as input (Jamf, Intune, etc).

A single entry can specify periods for both PARTIAL and FULL syncs, with the
latter taking precedence. In this case, it's expected for PARTIAL syncs to be
more frequent than FULL, so they are "squeezed" in-between FULL runs. (If
PARTIALs are less frequent than FULLs they would never run - in practice this
configuration is [not allowed][1].)

Each entry represents its own separate schedule and does not interfere in other
entries.

The scheduler itself is designed to be sleep free - the waiting is done by
whatever is using it, if desired. In our case, the specific MDM services will
code the wait using `time.Timer` or a similar mechanism.

#826

[1]: https://github.com/gravitational/teleport/blob/38bd6f00994efa983e436109c25755e702b135e2/api/types/jamf.go#L87-L89

Source-commit: gravitational/teleport.e@9b8185ca122db3f146ffaa31f6eb37b9c23d72a5
Source-pull-requests: gravitational/teleport.e#1554