[v13] Backport hardened AMI resources (#27454)
* Packer resources for hardened AMIs (#26126)

- Converts the original JSON Packer template to HCL, preserving the
   old json version for the legacy AMI builds
 - Updates and modernises the new HCL template
 - Creates a slimmed-down `install-hardened.sh` script that avoids
   installing the monitoring front-end tools, as these are a major
   source of CIS violations (e.g. pulling in X11)
 - Splits the exisiting resources in `files` into two classes:
   - `files` that is common to the old and new AMI, and
   - `monitor-files` that is only installed on the legacy AMI

Note that this patch does not include any high-level tools to drive this
packer script from CI/CD. That is coming in a later PR.

* Creates Hardened AMI promotion tool (#26754)

* Creates Hardened AMI promotion tool

Creates a tool that handles AMI promotion by finding the latest AMI
for a given Teleport release and makks it as public.

This patch also moves the `update-ami-id` script into its own package
directory in the `assets/aws` module, in order to both highlight its
use and tidy up the module root.

* Apply suggestions from code review

Co-authored-by: Noah Stride <noah.stride@goteleport.com>

* Update main.go

---------

Co-authored-by: Noah Stride <noah.stride@goteleport.com>

---------

Co-authored-by: Noah Stride <noah.stride@goteleport.com>