[v13] feat: label expressions (#27641)
* feat: partial label expression support (#26638)

* feat: partial label expression support

This commit implements support for label expressions (RFD 116) with the
exception that user traits can not yet be used in the expressions.
The user traits are made available to the expressions in a following
commit.

* feat: back compat improvements for label expressions (#27244)

- downgrade roles with label expressions for older clients
- add teleport.internal/downgraded label to downgraded roles
- block UpsertRole if inventory contains servers without label expression support

* feat: enable user traits in label expressions (#27138)

* chore: move accessChecker methods to access_checker.go (#27138)

* feat: extend label expression syntax (#27312)

This commit adds the new function `labels_matching` to the label
expression syntax. It accepts a Teleport-style glob or regular
expression, and returns the aggregate of all values for keys matching
that expression.

Since it is now more likely for users to want to compare two slices,
I've also added the `contains_any` and `contains_all` functions to
complement the existing `contains` function.

This commit also adds the `email.local` function that is already
available in role templates and should be useful.

Some examples are included in the new test cases.

* perf: benchmark label expressions (#27474)

* add benchmark
* performance improvements
* update RFD

* update minSupportedLabelExpressionVersion to v13.1.1

This feature will not actually work until api.Version >= minSupportedLabelExpressionVersion
v13.1.1 appears to be the next pending release, I will update if that
changes before this merges.